type conversion

+++ db/CategoryTable.sql
@@ -1,5 +1,5 @@
 create table CategoryTable(
-    categoryId int not null auto_increment,
+    categoryId BIGINT not null auto_increment,
     categoryName char(20) not null,
     primary key(categoryId)
 );
